added initial version of GDB support
It is of limited use as of now. Breakpoints cannot be set on jitted functions.
The disassembly command also does not work in generated code. You have to use
TUI mode.

A custom jit reader has to be used as asmjit does not generate ELF object
files which gdb could read. Build the shared library in the subdirectory
jit-reader with make and load it in gdb with jit-reader-load <path-to-so>.