added PerfCompiler which passes debug info through AsmJit's Compiler
It stores file name and source line internally and attaches a handle/index in
the user data of the instruction node. After the register allocator pass, when
serializing the assembly, we get the real offset in the machine code and add
this info to the jitdump file. All in all, perf report maps instructions back
to source lines which makes profiling much nicer.